Replaced all logging function with `warnings.warn()`
Replaced all loggers to default to warning.warn(). This makes it so that there are now excessing "warning" messages. These will need to be adjusted in the future version of TaxUMAP to make sure they are useful for the user.

Also, fixed the run_taxumap.py file so that it actually toggles saving on/off based on user input.